Combine custom-actions into action center
