+++ src/components/action-center/action-center.component.scss
@@ -15,6 +15,38 @@
   &.visible {
     transform: translateX(0);
   }
+
+  .custom-actions {
+    width: 30vw;
+    display: block;
+    overflow: scroll;
+
+    &::-webkit-scrollbar {
+      display: none;
+    }
+
+    &__action {
+      display: block;
+      float: left;
+      width: 10vw;
+      height: 10vw;
+      border: 0.8vw solid var(--border);
+      border-radius: 1.5vw;
+      text-align: center;
+      opacity: 0.9;
+      transition: opacity 0.2s;
+      margin: 0 1.4vw 3vh;
+
+      &[disabled='true'] {
+        opacity: 0.25;
+      }
+
+      &-icon {
+        font-size: 5.5vw;
+        vertical-align: -3vw;
+      }
+    }
+  }
 }
 
 .toggle__icon {
@@ -28,5 +60,83 @@
 
   &.visible {
     rotate: 180deg;
+  }
+}
+
+.iframe {
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  left: 0;
+  top: 0;
+  z-index: 100;
+  background: #2a2f39;
+  font-weight: 500;
+  font-size: 5vw;
+  text-align: center;
+  padding-top: 24vh;
+
+  &__content {
+    border: 0;
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+    position: fixed;
+    top: 0;
+    left: 0;
+  }
+
+  &__close {
+    width: 5vw;
+    position: fixed;
+    top: 3vh;
+    right: 3.5vw;
+    z-index: 101;
+    color: var(--error);
+  }
+}
+
+.confirm {
+  position: fixed;
+  top: 22vh;
+  width: 64vw;
+  margin-left: calc(15vw - 5px);
+  background-color: var(--background);
+  border: 5px solid rgb(255 255 255 / 40%);
+  border-radius: 2vw;
+  padding: 3vh 3vw;
+  z-index: 10;
+
+  &__sub-heading {
+    font-size: 2.4vw;
+    padding-top: 2vh;
+  }
+
+  &__gcode {
+    font-size: 2.2vw;
+    padding: 2vh 1.5vw;
+    font-family: Cousine, monospace;
+  }
+
+  &__icon {
+    font-size: 15vh;
+    height: 15vh;
+    margin-top: 7vh;
+    margin-left: 15vw;
+
+    &-yes {
+      color: var(--success);
+    }
+
+    &-no {
+      color: var(--error);
+    }
+  }
+
+  &__wrapper {
+    position: fixed;
+    z-index: 2;
+    background-color: rgb(0 0 0 / 85%);
+    inset: 0;
   }
 }
