--- src/components/action-center/custom-actions/custom-actions.component.html
@@ -1,38 +1,0 @@
-<div class="iframe" id="iframe" *ngIf="iframeOpen">
-  <span i18n="@@website-loading">loading website...</span>
-  <iframe
-    class="iframe__content"
-    [src]="iframeURL | url"
-    onerror="doActionConfirm()"
-    allow="fullscreen; camera;"></iframe>
-  <fa-icon icon="xmark" class="iframe__close" (click)="hideIframe()" />
-</div>
-
-<div class="custom-actions">
-  <div
-    class="custom-actions__action"
-    *ngFor="let customAction of customActions"
-    [attr.disabled]="!printerConnected && !customAction.command.startsWith('[!')"
-    [style.border-color]="customAction.color"
-    (click)="doAction(customAction.command, customAction.exit, customAction.confirm)"
-    matRipple
-    [matRippleUnbounded]="false">
-    <fa-icon
-      [icon]="customAction.icon"
-      class="custom-actions__action-icon"
-      [style.color]="customAction.color"></fa-icon>
-  </div>
-</div>
-
-<div class="confirm__wrapper" *ngIf="actionToConfirm">
-  <div class="confirm">
-    <h1 i18n="@@are-you-sure">are you sure?</h1>
-    <p class="confirm__sub-heading" i18n="@@confirm-exec">do you want to execute the following command?</p>
-    <p class="confirm__gcode">{{ actionToConfirm.command }}</p>
-    <fa-icon
-      icon="circle-check"
-      class="confirm__icon confirm__icon-yes"
-      (click)="doAction(actionToConfirm.command, actionToConfirm.exit, false)" />
-    <fa-icon icon="circle-xmark" class="confirm__icon confirm__icon-no" (click)="hideConfirm()" />
-  </div>
-</div>
