+++ src/components/control/control.component.html
@@ -135,7 +135,10 @@
     </div>
   </div>
 
-  <app-custom-actions [redirectActive]="true"></app-custom-actions>
+  <div class="control__tools">
+    <p>Temperature</p>
+    <!-- TODO: Place tool switching and bed temperature control here -->
+  </div>
 </div>
 
 <div class="quick-control" *ngIf="quickControlShown" id="quickControl">
